SetKitByName, hope it works
